db fix

+++ src/main/java/com/back/csaback/Repositories/EvaluationRepository.java
@@ -3,5 +3,5 @@ package com.back.csaback.Repositories;
 import com.back.csaback.Models.Evaluation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EvaluationRepository extends JpaRepository<Evaluation,Long> {
+public interface EvaluationRepository extends JpaRepository<Evaluation,Integer> {
 }
